update exception handle custom method

+++ src/main/java/edu/iuhs/crm/service/imp/StudentServiceImpl.java
@@ -2,11 +2,13 @@ package edu.iuhs.crm.service.imp;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import edu.iuhs.crm.entity.StudentEntity;
+import edu.iuhs.crm.exception.CustomNotDeleteStudentIdExeption;
+import edu.iuhs.crm.exception.CustomNotFoundException;
+import edu.iuhs.crm.exception.CustomNotFoundStudentsExceptionHandler;
 import edu.iuhs.crm.modle.Student;
 import edu.iuhs.crm.repository.StudentRepository;
 import edu.iuhs.crm.service.StudentService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -23,6 +25,9 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public Student persist(Student student) {
+        if (student.customNullValue(student)){
+            throw new CustomNotFoundException("Please fill the all data");
+        }
         StudentEntity studentEntity = studentRepository.save(
                 mapper.convertValue(student, StudentEntity.class)
         );
@@ -36,6 +41,8 @@ public class StudentServiceImpl implements StudentService {
     public List<Student> retrive() {
         List<Student> studentList = new ArrayList<>();
         Iterable<StudentEntity> allStudent = studentRepository.findAll();
+        if (allStudent==null)
+            throw new CustomNotFoundStudentsExceptionHandler("You can't view all Students");
         allStudent.forEach(studentEntity->{
             studentList.add(
                     mapper.convertValue(studentEntity, Student.class)
@@ -46,13 +53,23 @@ public class StudentServiceImpl implements StudentService {
 
     @Override
     public void delete(Integer id) {
-            studentRepository.deleteById(id);
+        Optional<StudentEntity> byId = studentRepository.findById(id);
+        if ( byId.isEmpty())
+            throw new CustomNotDeleteStudentIdExeption("You can't delete student");
+        studentRepository.deleteById(id);
     }
 
     @Override
     public Student update(Student student, Integer id) {
 
+        if (!student.customNullValue(student)) {
+            throw new CustomNotFoundException("Student Not Found");
+        }
+
         Optional<StudentEntity> byId = studentRepository.findById(id);
+        if ( byId.isEmpty())
+            throw new CustomNotDeleteStudentIdExeption("You can't update the student student");
+
         StudentEntity studentEntity = byId.get();
 
         studentEntity.setName(student.getName());
